+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 -   Fixed exponent in the expression of foreground count in p_astro_other
     task.
 
+-   Add ``PE_READY`` tag and upload link to PE results after PE finishes.
+
 0.2.6 (2019-02-12)
 ------------------